+++ gradle.properties
@@ -1,1 +1,1 @@
-version=1.0.0.BUILD-SNAPSHOT
+version=1.0.0.MARIUS-LOCAL
