+++ src/test/java/org/springframework/integration/kafka/listener/TestKafkaBrokerConnection.java
@@ -22,13 +22,14 @@ import org.junit.Assert;
 import org.junit.Rule;
 import org.junit.Test;
 
-import org.springframework.integration.kafka.serializer.common.StringDecoder;
 import org.springframework.integration.kafka.core.Connection;
-import org.springframework.integration.kafka.core.Result;
-import org.springframework.integration.kafka.core.Partition;
+import org.springframework.integration.kafka.core.DefaultConnection;
+import org.springframework.integration.kafka.core.FetchRequest;
 import org.springframework.integration.kafka.core.KafkaMessage;
 import org.springframework.integration.kafka.core.KafkaMessageBatch;
-import org.springframework.integration.kafka.core.FetchRequest;
+import org.springframework.integration.kafka.core.Partition;
+import org.springframework.integration.kafka.core.Result;
+import org.springframework.integration.kafka.serializer.common.StringDecoder;
 import org.springframework.integration.kafka.util.MessageUtils;
 
 /**
@@ -47,7 +48,7 @@ public class TestKafkaBrokerConnection extends AbstractBrokerTest {
 	@Test
 	public void testFetchPartitionMetadata() throws Exception {
 		createTopic(TEST_TOPIC, 1, 1, 1);
-		Connection brokerConnection = new Connection(getKafkaRule().getBrokerAddresses().get(0));
+		Connection brokerConnection = new DefaultConnection(getKafkaRule().getBrokerAddresses().get(0));
 		Partition partition = new Partition(TEST_TOPIC, 0);
 		Result<Long> result = brokerConnection.fetchInitialOffset(-1, partition);
 		Assert.assertEquals(0, result.getErrors().size());
@@ -60,7 +61,7 @@ public class TestKafkaBrokerConnection extends AbstractBrokerTest {
 		createTopic(TEST_TOPIC, 1, 1, 1);
 		Producer<String, String> producer = createStringProducer(0);
 		producer.send( createMessages(10));
-		Connection brokerConnection = new Connection(getKafkaRule().getBrokerAddresses().get(0));
+		Connection brokerConnection = new DefaultConnection(getKafkaRule().getBrokerAddresses().get(0));
 		Partition partition = new Partition(TEST_TOPIC, 0);
 		FetchRequest fetchRequest = new FetchRequest(partition, 0L, 1000);
 		Result<KafkaMessageBatch> result = brokerConnection.fetch(fetchRequest);
@@ -82,7 +83,7 @@ public class TestKafkaBrokerConnection extends AbstractBrokerTest {
 		createTopic(TEST_TOPIC, 1, 1, 1);
 		Producer<String, String> producer = createStringProducer(1);
 		producer.send( createMessages(10));
-		Connection brokerConnection = new Connection(getKafkaRule().getBrokerAddresses().get(0));
+		Connection brokerConnection = new DefaultConnection(getKafkaRule().getBrokerAddresses().get(0));
 		Partition partition = new Partition(TEST_TOPIC, 0);
 		FetchRequest fetchRequest = new FetchRequest(partition, 0L, 1000);
 		Result<KafkaMessageBatch> result = brokerConnection.fetch(fetchRequest);
@@ -104,7 +105,7 @@ public class TestKafkaBrokerConnection extends AbstractBrokerTest {
 		createTopic(TEST_TOPIC, 1, 1, 1);
 		Producer<String, String> producer = createStringProducer(2);
 		producer.send( createMessages(10));
-		Connection brokerConnection = new Connection(getKafkaRule().getBrokerAddresses().get(0));
+		Connection brokerConnection = new DefaultConnection(getKafkaRule().getBrokerAddresses().get(0));
 		Partition partition = new Partition(TEST_TOPIC, 0);
 		FetchRequest fetchRequest = new FetchRequest(partition, 0L, 1000);
 		Result<KafkaMessageBatch> result = brokerConnection.fetch(fetchRequest);
