+++ src/test/java/org/springframework/integration/kafka/listener/TestKafkaInboundChannelAdapter.java
@@ -20,9 +20,6 @@ package org.springframework.integration.kafka.listener;
 import static org.hamcrest.collection.IsCollectionWithSize.hasSize;
 import static org.hamcrest.core.IsEqual.equalTo;
 import static org.junit.Assert.assertThat;
-import static org.springframework.integration.kafka.inbound.KafkaInboundChannelAdapter.KAFKA_MESSAGE_KEY;
-import static org.springframework.integration.kafka.inbound.KafkaInboundChannelAdapter.KAFKA_MESSAGE_OFFSET;
-import static org.springframework.integration.kafka.inbound.KafkaInboundChannelAdapter.KAFKA_MESSAGE_PARTITION;
 
 import java.util.ArrayList;
 import java.util.concurrent.CountDownLatch;
@@ -38,6 +35,7 @@ import org.springframework.integration.kafka.core.ConnectionFactory;
 import org.springframework.integration.kafka.core.Partition;
 import org.springframework.integration.kafka.inbound.KafkaInboundChannelAdapter;
 import org.springframework.integration.kafka.serializer.common.StringDecoder;
+import org.springframework.integration.kafka.support.KafkaHeaders;
 import org.springframework.messaging.Message;
 import org.springframework.messaging.MessageChannel;
 
@@ -83,13 +81,13 @@ public class TestKafkaInboundChannelAdapter extends AbstractMessageListenerConta
 			public boolean send(Message<?> message) {
 				latch.countDown();
 				return receivedData.put(
-						(Integer)message.getHeaders().get(KAFKA_MESSAGE_PARTITION),
+						(Integer)message.getHeaders().get(KafkaHeaders.PARTITION_ID),
 						new KeyedMessageWithOffset(
-								(String)message.getHeaders().get(KAFKA_MESSAGE_KEY),
+								(String)message.getHeaders().get(KafkaHeaders.MESSAGE_KEY),
 								(String)message.getPayload(),
-								(Long)message.getHeaders().get(KAFKA_MESSAGE_OFFSET),
+								(Long)message.getHeaders().get(KafkaHeaders.OFFSET),
 								Thread.currentThread().getName(),
-								(Integer)message.getHeaders().get(KAFKA_MESSAGE_PARTITION)));
+								(Integer)message.getHeaders().get(KafkaHeaders.PARTITION_ID)));
 			}
