+++ src/test/java/org/springframework/integration/kafka/listener/TestMultiBroker5ReplicatedWithBounce.java
@@ -35,6 +35,7 @@ import org.junit.Rule;
 import org.junit.Test;
 
 import org.springframework.integration.kafka.core.ConnectionFactory;
+import org.springframework.integration.kafka.core.DefaultConnectionFactory;
 import org.springframework.integration.kafka.core.KafkaMessage;
 import org.springframework.integration.kafka.core.Partition;
